// public-report-client/src/api/lost-items/model/createLostReportRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ItemAttributesDto } from './itemAttributesDto';

export interface CreateLostReportRequest {
  /**
   * @minLength 0
   * @maxLength 2000
   */
  description: string;
  lostAt: string;
  /**
   * @minLength 0
   * @maxLength 255
   */
  location?: string;
  venueId?: string;
  /**
   * @minLength 0
   * @maxLength 255
   */
  contactEmail: string;
  attributes?: ItemAttributesDto;
}